+++ tools/sampcd_processor_utils.py
@@ -265,7 +265,7 @@ def extract_code_blocks_from_docstr(docstr, google_style=True):
     Return:
         code_blocks: A list of code-blocks, indent removed.
                      element {'name': the code-block's name, 'id': sequence id.
-                              'codes': codes, 'required': 'gpu', 'in_examples': bool, code block in `Examples` or not,}
+                              'codes': codes, 'in_examples': bool, code block in `Examples` or not,}
     """
     code_blocks = []
 
@@ -290,7 +290,6 @@ def extract_code_blocks_from_docstr(docstr, google_style=True):
 
     cb_start_pat = re.compile(r"code-block::\s*python")
     cb_param_pat = re.compile(r"^\s*:(\w+):\s*(\S*)\s*$")
-    cb_required_pat = re.compile(r"^\s*#\s*require[s|d]\s*:\s*(\S+)\s*$")
 
     cb_info = {}
     cb_info['cb_started'] = False
@@ -298,23 +297,20 @@ def extract_code_blocks_from_docstr(docstr, google_style=True):
     cb_info['cb_cur_indent'] = -1
     cb_info['cb_cur_name'] = None
     cb_info['cb_cur_seq_id'] = 0
-    cb_info['cb_required'] = None
 
     def _cb_started():
-        # nonlocal cb_started, cb_cur_name, cb_required, cb_cur_seq_id
+        # nonlocal cb_started, cb_cur_name, cb_cur_seq_id
         cb_info['cb_started'] = True
         cb_info['cb_cur_seq_id'] += 1
         cb_info['cb_cur_name'] = None
-        cb_info['cb_required'] = None
 
     def _append_code_block(in_examples):
-        # nonlocal code_blocks, cb_cur, cb_cur_name, cb_cur_seq_id, cb_required
+        # nonlocal code_blocks, cb_cur, cb_cur_name, cb_cur_seq_id
         code_blocks.append(
             {
                 'codes': inspect.cleandoc("\n" + "\n".join(cb_info['cb_cur'])),
                 'name': cb_info['cb_cur_name'],
                 'id': cb_info['cb_cur_seq_id'],
-                'required': cb_info['cb_required'],
                 'in_examples': in_examples,
             }
         )
@@ -339,10 +335,6 @@ def extract_code_blocks_from_docstr(docstr, google_style=True):
                     if mo_p.group(1) == 'name':
                         cb_info['cb_cur_name'] = mo_p.group(2)
                     continue
-                # read the required directive
-                mo_r = cb_required_pat.match(linecont)
-                if mo_r:
-                    cb_info['cb_required'] = mo_r.group(1)
                 # docstring end
                 if lineno == lastlineindex:
                     mo = re.search(r"\S", linecont)
@@ -479,6 +471,49 @@ def get_docstring(full_test=False):
     return docstrings_to_test, whl_error
 
 
+def check_old_style(docstrings_to_test: typing.Dict[str, str]):
+    old_style_apis = []
+    for api_name, raw_docstring in docstrings_to_test.items():
+        for codeblock in extract_code_blocks_from_docstr(
+            raw_docstring, google_style=False
+        ):
+            old_style = True
+
+            for line in codeblock['codes'].splitlines():
+                if line.strip().startswith('>>>'):
+                    old_style = False
+                    break
+
+            if old_style:
+                codeblock_name = codeblock['name']
+                codeblock_id = codeblock['id']
+
+                docstring_name = '{}:{}'.format(
+                    api_name, codeblock_name or codeblock_id
+                )
+
+                old_style_apis.append(docstring_name)
+
+    if old_style_apis:
+        stdout_handler = logging.StreamHandler(stream=sys.stdout)
+        logger.addHandler(stdout_handler)
+
+        logger.info(
+            ">>> %d apis use plain sample code style.",
+            len(old_style_apis),
+        )
+        logger.info('=======================')
+        logger.info('\n'.join(old_style_apis))
+        logger.info('=======================')
+        logger.info("Check Failed!")
+        logger.info("DEPRECATION: Please do not use plain sample code style.")
+        logger.info(
+            "For more information: https://www.paddlepaddle.org.cn/documentation/docs/zh/develop/dev_guides/style_guide_and_references/code_example_writing_specification_cn.html "
+        )
+        logger.info("----------------End of the Check--------------------")
+        sys.exit(1)
+
+
 def exec_gen_doc():
     result = True
     cmd = ["bash", "document_preview.sh"]
@@ -568,27 +603,30 @@ def get_test_results(
 
 
 def run_doctest(args, doctester: DocTester):
-    logger.info("----------------Codeblock Check Start--------------------")
-
     # init logger
     init_logger(debug=args.debug, log_file=args.logf)
 
-    logger.info("Check test mode ...")
+    logger.info("----------------Codeblock Check Start--------------------")
+
+    logger.info(">>> Check test mode ...")
     run_on_device = check_test_mode(mode=args.mode, gpu_id=args.gpu_id)
 
-    logger.info("Get test capacity ...")
+    logger.info(">>> Get test capacity ...")
     sample_code_test_capacity = get_test_capacity(run_on_device)
 
-    logger.info("Get docstring from api ...")
+    logger.info(">>> Get docstring from api ...")
     docstrings_to_test, whl_error = get_docstring(full_test=args.full_test)
 
-    logger.info("Prepare doctester ...")
+    logger.info(">>> Checking plain sample code style before Paddle 2.5 ...")
+    check_old_style(docstrings_to_test)
+
+    logger.info(">>> Prepare doctester ...")
     doctester.prepare(sample_code_test_capacity)
 
-    logger.info("Running doctester ...")
+    logger.info(">>> Running doctester ...")
     test_results = get_test_results(doctester, docstrings_to_test)
 
-    logger.info("Print summary ...")
+    logger.info(">>> Print summary ...")
     doctester.print_summary(test_results, whl_error)
 
     if args.mode == "cpu":
@@ -596,37 +634,39 @@ def run_doctest(args, doctester: DocTester):
         exec_gen_doc()
 
 
-arguments = [
-    # flags, dest, type, default, help
-    ['--gpu_id', 'gpu_id', int, 0, 'GPU device id to use [0]'],
-    ['--logf', 'logf', str, None, 'file for logging'],
-    ['--threads', 'threads', int, 0, 'sub processes number'],
-]
-
-
 def parse_args():
     """
     Parse input arguments
     """
-    global arguments
     parser = argparse.ArgumentParser(description='run Sample Code Test')
     parser.add_argument('--debug', dest='debug', action="store_true")
     parser.add_argument('--full-test', dest='full_test', action="store_true")
-    parser.add_argument('mode', type=str, help='run on device', default='cpu')
+    parser.add_argument(
+        '--mode', dest='mode', type=str, default='cpu', help='run on device'
+    )
     parser.add_argument(
         '--build-doc',
         dest='build_doc',
         action='store_true',
         help='build doc if need.',
     )
-    for item in arguments:
-        parser.add_argument(
-            item[0], dest=item[1], help=item[4], type=item[2], default=item[3]
-        )
-
-    if len(sys.argv) == 1:
-        args = parser.parse_args(['cpu'])
-        return args
+    parser.add_argument(
+        '--gpu_id',
+        dest='gpu_id',
+        type=int,
+        default=0,
+        help='GPU device id to use [0]',
+    )
+    parser.add_argument(
+        '--logf', dest='logf', type=str, default=None, help='file for logging'
+    )
+    parser.add_argument(
+        '--threads',
+        dest='threads',
+        type=int,
+        default=0,
+        help='sub processes number',
+    )
 
     args = parser.parse_args()
     return args
